requirements.txt
alembic==1.14.1
click==8.1.7
flask_api==3.1
flask_mail==0.10.0
flask_migrate==4.1.0
flask_sqlalchemy==3.1.1
itsdangerous==2.2.0
PyJWT==2.8.0
PyJWT==2.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
Werkzeug==3.1.3


